+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - slight change for indirect parameter order in processes (backward compatible)
 
 ### Fixed
+- fixed #46 were the process link of arguments was not set correctly
 - fixed the server sided validation response handling (HTTP 200 were always OK, despite having errors)
 - fixed bad ordering of bounding box parameter